+++ CHANGELOG.md
@@ -10,9 +10,11 @@
 - Added `measureText()` two function for computing width and height of `TwoText` objects without rendering.
 - Added a search suggestion autocomplete feature.
 - Added a feature to request for bind TFs in a certain interval using Cistrome DB API and render the result in a simple data table.
+- Added a context menu for filtering or highlighting rows based on a nomival value in categorical bar charts.
 
 ### Changed
 - Fixed bug which prevented wrapper options for sorting, filtering, highlighting from being used upon initial component render.
+- Added resizer elements to their containers and used `d3-drag` for the mouse events.
 
 
 ## 0.1.0 - 02/19/20